gpt: use 64-bit sector arithmetic and harden entry parsing
GPT LBAs are 64-bit but were narrowed to unsigned int, so a partition
beyond 2 TB wrapped and a name-addressed write/erase hit the wrong
sectors. Carry start_sector/num_sectors as uint64_t and range-check
before storing num_sectors. Also reject a bad part_entry_size, fix an
off-by-one in the UTF-16 surrogate handling that read past the name
buffer, and initialise the sim-mode outputs.

Fixes: 60db50966d4e ("qdl: Extend read/write support to accept GPT partition names")

+++ src/gpt.h
@@ -2,13 +2,15 @@
 #ifndef __GPT_H__
 #define __GPT_H__
 
+#include <stdint.h>
+
 #include "list.h"
 
 struct qdl_device;
 struct list_head;
 
 int gpt_find_by_name(struct qdl_device *qdl, const char *name, int *partition,
-		     unsigned int *start_sector, unsigned int *num_sectors);
+		     uint64_t *start_sector, uint64_t *num_sectors);
 int gpt_resolve_deferrals(struct qdl_device *qdl, struct list_head *ops);
 
 #endif
